// src/enums/bot-commands.enum.ts
export enum BotCommands {
  Help = '!help',
  Time = '!time',
  Date = '!date',
  Facebook = '!facebook',
  Twitter = '!twitter',
  Youtube = '!youtube',
  Twitch = '!twitch',
  Instagram = '!instagram',
  Discord = '!discord',
  Invite = '!invite',
  Email = '!email',
  Git = '!git',
  Github = '!github',
  Welcome = '!welcome',
  Warn = '!warn',
  Rules = '!rules',
  Kick = '!kick',
  Ban = '!ban',
  P2C = '!p2c',
  // CreateChannel = '!createChannel',
  // DeleteChannel = '!delChannel',
  DeleteHistory = '!delhistory',
  Hi = '!hi',
  Hello = '!hello',
  Bye = '!bye',
  Imsak = '!imsak',
  Sunrise = '!sunrise',
  Fajr = '!fajr',
  Dhuhr = '!dhuhr',
  Asr = '!asr',
  Iftar = '!iftar',
  Maghrib = '!maghrib',
  Isha = '!isha',
  Midnight = '!midnight'
}